Logic Gate Programmer
This application is used to deploy the standard gate configurations to the BreadboarD GeniuS Programmable logic gates.

You need to download the supporting .hex files and place them in the same directory as the programmer:
and-nand.hex
or-nor.hex
xor-xnor.hex
dual-not.hex
majority-minority.hex
LEDFlash.hex

Python3 and Arduino IDE with MegaTinyCore need to be installed for the application to function.

Additional custom .hex files can be loaded into the application for quick deployment, those can be written and compiled within Arudino IDE.

The reason for the application is to create a quick interface for configuring the BreadboarD GeniuS Logic gates. The logic gate kit comes with a programming module built into the case that stores the logic gates, you can quickly remove a logic gate, place it in the programmer which is connected to your Windows PC via USB, then simply click the logic type your gate is to be programmed with. each logic gate has a changeable identification cap with the logic symbol, the Logic gate kit comes with 22 programmable logic gates and 20 additional caps of each logic type. if you need 22 AND/NAND gates then you can program all the logic gates as AND/NAND, then place the identification cap on the top.